Ensure replicas are actually replicating for cluster-health
It could be that a replica is not replicating properly from the leader,
however, this was not directly caught so far (only when the lag was getting too
large).  To improve the cluster-health check, get the leader's `member' API
response and check how many rows the `replication' object has. If that number
is different to the number of replicas, consider the cluster as being
unhealthy.

In passing, also log the reason for the failing cluster-health at warning
level.